+++ picoserve/src/sync/oneshot_broadcast.rs
@@ -34,40 +34,28 @@ impl<'a, T: Copy> Signal<'a, T> {
 
     pub fn listen(&self) -> Listener<'a, T> {
         Listener {
-            channel: Some(self.channel),
+            channel: self.channel,
         }
     }
 }
 
 #[derive(Clone)]
 pub struct Listener<'a, T: Copy> {
-    channel: Option<&'a SignalCore<T>>,
-}
-
-impl<T: Copy> Listener<'_, T> {
-    pub fn never() -> Self {
-        Self { channel: None }
-    }
+    channel: &'a SignalCore<T>,
 }
 
 impl<T: Copy> core::future::Future for Listener<'_, T> {
     type Output = T;
 
     fn poll(
-        mut self: core::pin::Pin<&mut Self>,
+        self: core::pin::Pin<&mut Self>,
         cx: &mut core::task::Context<'_>,
     ) -> core::task::Poll<Self::Output> {
-        let Some(mux) = self.channel else {
-            return core::task::Poll::Pending;
-        };
-
-        if let Some(value) = mux.value.get() {
-            self.channel = None;
-
+        if let Some(value) = self.channel.value.get() {
             return core::task::Poll::Ready(value);
         }
 
-        let new_waker = if let Some(current_waker) = mux.waker.take() {
+        let new_waker = if let Some(current_waker) = self.channel.waker.take() {
             if current_waker.will_wake(cx.waker()) {
                 current_waker
             } else {
@@ -79,7 +67,7 @@ impl<T: Copy> core::future::Future for Listener<'_, T> {
             cx.waker().clone()
         };
 
-        mux.waker.set(Some(new_waker));
+        self.channel.waker.set(Some(new_waker));
 
         core::task::Poll::Pending
     }